requirements.txt
espn-api==0.45.1
openai==1.3.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
plotly==5.17.0
dash==2.14.2
schedule==1.2.0
beautifulsoup4==4.12.2
lxml==4.9.3
